# apps/functions/lib/swagger.yaml
openapi: 3.0.0
info:
  title: SAE Group Upfitter OS API
  description: Professional documentation for the SAE Group Operational API. Covers Jobs, Tasks, Daily Logs, CRM, Vehicles, Tools, HR, Inventory, and Sales.
  version: 1.0.0
servers:
  - url: http://localhost:5001/saegroup-os/us-central1/api
    description: Local emulator
  - url: https://us-central1-saegroup-os.cloudfunctions.net/api
    description: Production API
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Area:
      type: object
      properties:
        id: { type: string }
        tenantId: { type: string }
        type: { type: string }
        label: { type: string }
        status: { type: string }
        capacity: { type: string }
        wallHeight: { type: number }
        color: { type: string }
        points:
          type: array
          items:
            type: object
            properties:
              x: { type: number }
              y: { type: number }
...
security:
  - BearerAuth: []
paths:
...
  /areas:
    get:
      summary: List all Areas and Map Zones
      tags: [Areas & Facilities]
      parameters:
        - in: query
          name: tenantId
          schema:
            type: string
          required: true
      responses:
        '200':
          description: Array of Areas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Area'
    post:
      summary: Register a new Area
      tags: [Areas & Facilities]
      responses:
        '201':
          description: Area Registered
  /areas/{id}:
    put:
      summary: Update Area Details
      tags: [Areas & Facilities]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Updated
    delete:
      summary: Delete an Area
      tags: [Areas & Facilities]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Deleted
...