# apps/functions/src/routes/areas.routes.ts
import { Router, Request, Response } from 'express';
import * as admin from 'firebase-admin';
import { authenticate } from '../middleware/auth.middleware';

export const areasRoutes = Router();
const getDb = () => admin.firestore();

// Helper to determine if the caller has access to the workspace
const isMemberOfTenant = (caller: any, tenantId: string) => {
    const isSuperAdmin = (caller.role === 'system_owner' || caller.role === 'super_admin');
    const isTenantMember = caller.tenantId === tenantId;
    return isSuperAdmin || isTenantMember;
};

const isManagerOfTenant = (caller: any, tenantId: string) => {
    const isSuperAdmin = (caller.role === 'system_owner' || caller.role === 'super_admin');
    const isTenantManager = (caller.role === 'business_owner' || caller.role === 'manager') && caller.tenantId === tenantId;
    return isSuperAdmin || isTenantManager;
};

// GET /areas?tenantId=xyz - Fetch areas mapped from business_zones
areasRoutes.get('/', authenticate, async (req: Request, res: Response): Promise<Response> => {
    try {
        const caller = (req as any).user;
        const tenantId = req.query.tenantId as string;

        if (!tenantId) {
            return res.status(400).json({ error: 'tenantId query parameter is required' });
        }

        if (!isMemberOfTenant(caller, tenantId)) {
            return res.status(403).json({ error: 'Forbidden. You do not have access to this workspace.' });
        }

        const snapshot = await getDb().collection('business_zones')
            .where('tenantId', '==', tenantId)
            .get();

        let areas = snapshot.docs.map(doc => ({ id: doc.id, ...doc.data() as any }));
        
        areas.sort((a, b) => {
            const timeA = new Date(a.createdAt || 0).getTime();
            const timeB = new Date(b.createdAt || 0).getTime();
            return timeB - timeA;
        });

        return res.json(areas);
    } catch (error) {
        console.error("Error fetching areas from zones view:", error);
        return res.status(500).json({ error: 'Failed to fetch areas' });
    }
});

// POST /areas - Create a new area as an unmapped zone that can be placed later
areasRoutes.post('/', authenticate, async (req: Request, res: Response): Promise<Response> => {
    try {
        const caller = (req as any).user;
        const tenantId = req.body.tenantId;

        if (!tenantId) {
            return res.status(400).json({ error: 'tenantId is required' });
        }

        if (!isMemberOfTenant(caller, tenantId)) {
            return res.status(403).json({ error: 'Forbidden. You do not have access to this workspace.' });
        }

        const {
            label, type, color, status, capacity, notes, floorId, wallHeight
        } = req.body;

        const newId = `zone_${Date.now()}`;
        
        const timestamp = new Date().toISOString();
        const activityLogEntry = {
            timestamp,
            actor: caller.displayName || caller.email || 'Unknown Staff',
            action: `Created area record`
        };

        const newArea = {
            id: newId,
            tenantId,
            type: type || 'Bay',
            color: color || '#3b82f6',
            label: label || 'New Unassigned Area',
            wallHeight: wallHeight || 10,
            floorId: floorId || 'default',
            status: status || 'Active',
            capacity: capacity || '',
            notes: notes || '',
            points: [], // Implicitly unmapped geometrically
            width: 0,
            height: 0,
            createdBy: caller.displayName || caller.email || 'Unknown Staff',
            lastModifiedBy: caller.displayName || caller.email || 'Unknown Staff',
            createdAt: timestamp,
            updatedAt: timestamp,
        };

        await getDb().collection('business_zones').doc(newId).set({
            ...newArea,
            activityLogs: admin.firestore.FieldValue.arrayUnion(activityLogEntry)
        });
        
        return res.status(201).json({
            ...newArea,
            activityLogs: [activityLogEntry]
        });
    } catch (error) {
        console.error("Error creating area mapped to zone:", error);
        return res.status(500).json({ error: 'Failed to create area record' });
    }
});

// PUT /areas/:id - Update an area
areasRoutes.put('/:id', authenticate, async (req: Request, res: Response): Promise<Response> => {
    try {
        const caller = (req as any).user;
        const areaId = req.params.id;

        const areaRef = getDb().collection('business_zones').doc(areaId);
        const areaDoc = await areaRef.get();

        if (!areaDoc.exists) {
            return res.status(404).json({ error: 'Area not found' });
        }

        const areaData = areaDoc.data()!;
        const tenantId = areaData.tenantId;

        if (!isMemberOfTenant(caller, tenantId)) {
            return res.status(403).json({ error: 'Forbidden. Cannot update this area.' });
        }

        const {
            label, type, color, status, capacity, notes, wallHeight, floorId
        } = req.body;

        const updates: any = { 
            updatedAt: new Date().toISOString(),
            lastModifiedBy: caller.displayName || caller.email || 'Unknown Staff',
        };

        if (label !== undefined) {
            updates.label = label;
            updates.activityLogs = admin.firestore.FieldValue.arrayUnion({
                timestamp: new Date().toISOString(),
                actor: caller.displayName || caller.email || 'Unknown Staff',
                action: `Updated label`
            });
        }
        if (type !== undefined) updates.type = type;
        if (color !== undefined) updates.color = color;
        if (status !== undefined) updates.status = status;
        if (capacity !== undefined) updates.capacity = capacity;
        if (notes !== undefined) updates.notes = notes;
        if (wallHeight !== undefined) updates.wallHeight = wallHeight;
        if (floorId !== undefined) updates.floorId = floorId;

        await areaRef.update(updates);
        
        // Strip out the FieldValue from the updates we return to the client
        const { activityLogs, ...safeUpdates } = updates;
        
        return res.json({ id: areaId, ...areaData, ...safeUpdates });
    } catch (error) {
        console.error("Error updating area in zones db:", error);
        return res.status(500).json({ error: 'Failed to update area' });
    }
});

// DELETE /areas/:id - Delete an area/zone
areasRoutes.delete('/:id', authenticate, async (req: Request, res: Response): Promise<Response> => {
    try {
        const caller = (req as any).user;
        const areaId = req.params.id;

        const areaRef = getDb().collection('business_zones').doc(areaId);
        const areaDoc = await areaRef.get();

        if (!areaDoc.exists) {
            return res.status(404).json({ error: 'Area not found' });
        }

        const areaData = areaDoc.data()!;
        const tenantId = areaData.tenantId;

        if (!isManagerOfTenant(caller, tenantId)) {
            return res.status(403).json({ error: 'Forbidden. Only managers can delete areas.' });
        }

        await areaRef.delete();
        return res.json({ message: 'Area deleted successfully from active map' });
    } catch (error) {
        console.error("Error deleting area:", error);
        return res.status(500).json({ error: 'Failed to delete area' });
    }
});